+++ cmd/gen/workflows/runner.go
@@ -47,7 +47,8 @@ func (r *runner) run(ctx context.Context, _ *cobra.Command, _ []string) error {
 	var workflowsInput *workflows.Workflows
 	{
 		c := workflows.Config{
-			Flavours: r.flag.Flavours,
+			Flavours:                  r.flag.Flavours,
+			ReleaseCandidateByDefault: r.flag.ReleaseCandidateByDefault,
 		}
 
 		workflowsInput, err = workflows.New(c)